Disable GUI and NFS sharing; define vm hostname
Most of the challenges (if not all) will be resolved using the terminal,
disable the GUI.

Vagrant warns about using shared folders with untrusted boxes,
prefer disabling it and also, reduce the chance to be used as
cheat.

Define the hostname of the vm so it does not say 'vagrant' in
the prompt. Eventually we will be using multiples vms and the
hostname will be a hint for the student that it may be in the
wrong vm.